moved cropping.py to the operations module
