+++ pam/operations/cropping.py
